+++ catalyst_sdwan_lab/tasks/restore.py
@@ -23,6 +23,7 @@ from . import delete
 from .utils import (
     DATA_DIR,
     check_manager_ip_is_free,
+    complete_initial_setup_workflow,
     configure_manager_basic_settings,
     get_cml_sdwan_image_definition,
     get_sdwan_lab_parameters,
@@ -263,7 +264,6 @@ def main(
             existing_manager_users = re.findall(
                 r"<user>[\s\S]+?<name>(\w+)</name>", manager_node["configuration"]
             )
-            print(existing_manager_users)
             existing_manager_users.remove("admin")
             if existing_manager_users and manager_user not in existing_manager_users:
                 # Use regex to replace only the username within <name> tags to avoid
@@ -382,10 +382,11 @@ def main(
     manager_node.wait_until_converged()
     # Wait for SD-WAN Manager API to be available
     manager_session = wait_for_manager_session(
-        manager_ip, manager_port, manager_user, manager_password, log
+        manager_ip, manager_port, manager_user, manager_password, software_version, log
     )
     # Configure basic settings like org-name, validator fqdn etc.
     configure_manager_basic_settings(manager_session, ca_chain, log)
+    complete_initial_setup_workflow(manager_session, software_version, log)
 
     # Add controllers to SD-WAN Manager and sing certificates
     onboard_control_components(